+++ src/SMAPI.Installer/InteractiveInstaller.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Globalization;
 using System.IO;
 using System.Linq;
-using System.Reflection;
 using System.Threading;
 using Microsoft.Win32;
 using StardewModdingApi.Installer.Enums;
@@ -22,8 +21,8 @@ namespace StardewModdingApi.Installer
         /*********
         ** Properties
         *********/
-        /// <summary>The name of the installer file in the package.</summary>
-        private readonly string InstallerFileName = "install.exe";
+        /// <summary>The absolute path to the directory containing the files to copy into the game folder.</summary>
+        private readonly string BundlePath;
 
         /// <summary>The <see cref="Environment.OSVersion"/> value that represents Windows 7.</summary>
         private readonly Version Windows7Version = new Version(6, 1);
@@ -153,8 +152,10 @@ namespace StardewModdingApi.Installer
         ** Public methods
         *********/
         /// <summary>Construct an instance.</summary>
-        public InteractiveInstaller()
+        /// <param name="bundlePath">The absolute path to the directory containing the files to copy into the game folder.</param>
+        public InteractiveInstaller(string bundlePath)
         {
+            this.BundlePath = bundlePath;
             this.ConsoleWriter = new ColorfulConsoleWriter(EnvironmentUtility.DetectPlatform(), MonitorColorScheme.AutoDetect);
         }
 
@@ -329,8 +330,8 @@ namespace StardewModdingApi.Installer
                 }
 
                 // get folders
-                DirectoryInfo packageDir = new DirectoryInfo(Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location));
-                paths = new InstallerPaths(packageDir, installDir, EnvironmentUtility.GetExecutableName(platform));
+                DirectoryInfo bundleDir = new DirectoryInfo(this.BundlePath);
+                paths = new InstallerPaths(bundleDir, installDir, EnvironmentUtility.GetExecutableName(platform));
             }
             Console.Clear();
 
@@ -338,23 +339,11 @@ namespace StardewModdingApi.Installer
             /*********
             ** Step 4: validate assumptions
             *********/
+            if (!File.Exists(paths.ExecutablePath))
             {
-                if (!paths.PackageDir.Exists)
-                {
-                    this.PrintError(platform == Platform.Windows && paths.PackagePath.Contains(Path.GetTempPath()) && paths.PackagePath.Contains(".zip")
-                        ? "The installer is missing some files. It looks like you're running the installer from inside the downloaded zip; make sure you unzip the downloaded file first, then run the installer from the unzipped folder."
-                        : $"The 'internal/{paths.PackageDir.Name}' package folder is missing (should be at {paths.PackagePath})."
-                    );
-                    Console.ReadLine();
-                    return;
-                }
-
-                if (!File.Exists(paths.ExecutablePath))
-                {
-                    this.PrintError("The detected game install path doesn't contain a Stardew Valley executable.");
-                    Console.ReadLine();
-                    return;
-                }
+                this.PrintError("The detected game install path doesn't contain a Stardew Valley executable.");
+                Console.ReadLine();
+                return;
             }
 
 
@@ -446,11 +435,8 @@ namespace StardewModdingApi.Installer
                 {
                     // copy SMAPI files to game dir
                     this.PrintDebug("Adding SMAPI files...");
-                    foreach (FileSystemInfo sourceEntry in paths.PackageDir.EnumerateFileSystemInfos().Where(this.ShouldCopy))
+                    foreach (FileSystemInfo sourceEntry in paths.BundleDir.EnumerateFileSystemInfos().Where(this.ShouldCopy))
                     {
-                        if (sourceEntry.Name.StartsWith(this.InstallerFileName)) // e.g. install.exe or install.exe.config
-                            continue;
-
                         this.InteractivelyDelete(Path.Combine(paths.GameDir.FullName, sourceEntry.Name));
                         this.RecursiveCopy(sourceEntry, paths.GameDir);
                     }
@@ -478,14 +464,14 @@ namespace StardewModdingApi.Installer
                     }
 
                     // add or replace bundled mods
-                    DirectoryInfo packagedModsDir = new DirectoryInfo(Path.Combine(paths.PackageDir.FullName, "Mods"));
-                    if (packagedModsDir.Exists && packagedModsDir.EnumerateDirectories().Any())
+                    DirectoryInfo bundledModsDir = new DirectoryInfo(Path.Combine(paths.BundlePath, "Mods"));
+                    if (bundledModsDir.Exists && bundledModsDir.EnumerateDirectories().Any())
                     {
                         this.PrintDebug("Adding bundled mods...");
 
                         ModToolkit toolkit = new ModToolkit();
                         ModFolder[] targetMods = toolkit.GetModFolders(paths.ModsPath).ToArray();
-                        foreach (ModFolder sourceMod in toolkit.GetModFolders(packagedModsDir.FullName))
+                        foreach (ModFolder sourceMod in toolkit.GetModFolders(bundledModsDir.FullName))
                         {
                             // validate source mod
                             if (sourceMod.Manifest == null)
@@ -527,7 +513,7 @@ namespace StardewModdingApi.Installer
                     }
 
                     // remove obsolete appdata mods
-                    this.InteractivelyRemoveAppDataMods(paths.ModsDir, packagedModsDir);
+                    this.InteractivelyRemoveAppDataMods(paths.ModsDir, bundledModsDir);
                 }
             }
             Console.WriteLine();